Implemented TlsmTableFind(); fixed quite a few bugs in merge, comparators, etc...

+++ stasis/common.h
@@ -100,9 +100,9 @@ extern int errno;
 typedef unsigned char byte;
 //@todo lsn_t should be unsigned.
 // If it were unsigned, it could be typedef'ed from size_t.
-typedef long long lsn_t;
+typedef int64_t lsn_t;
 #define LSN_T_MAX    INT64_MAX
-typedef long long pageid_t;
+typedef int64_t pageid_t;
 #define PAGEID_T_MAX INT64_MAX
 
 /*#define DEBUGGING   */
